AI-NLP/text-classifier/requirements.txt
appdirs==1.4.4
APScheduler==3.9.1.post1
asgiref==3.4.1
astroid==2.6.6
atomicwrites==1.4.0
attrs==21.2.0
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
blis==0.7.4
boto3==1.24.58
botocore==1.27.96
bson==0.5.10
catalogue==1.0.0
certifi==2021.5.30
cfgv==3.3.1
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
cymem==2.0.5
distlib==0.3.3
Django==3.2.5
django-allow-cidr==0.4.0
django-apscheduler==0.6.0
django-cors-headers==3.8.0
django-environ==0.8.1
djangorestframework==3.12.4
filelock==3.1.0
greenlet==2.0.1
gunicorn==20.1.0
huggingface-hub==0.0.12
identify==2.2.15
idna==3.2
importlib-metadata==1.3.0
iniconfig==1.1.1
jmespath==1.0.1
joblib==1.2.0
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
murmurhash==1.0.5
mypy-extensions==0.4.3
mysqlclient==2.1.0
netaddr==0.8.0
nltk==3.6.2
numpy==1.21.2
packaging==21.0
pandas==1.3.3
pathspec==0.9.0
pathy==0.6.0
plac==1.1.3
platformdirs==2.4.0
pluggy==0.13.1
preshed==3.0.5
py==1.10.0
pydantic==1.8.2
PyJWT==2.2.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.2
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
regex==2021.8.28
requests==2.26.0
s3transfer==0.6.0
sacremoses==0.0.53
scikit-learn==0.24.2
scipy==1.7.3
six==1.16.0
sklearn==0.0.post1
smart-open==5.2.1
spacy==2.3.5
spacy-legacy==3.0.8
SQLAlchemy==1.4.44
sqlparse==0.4.2
srsly==1.0.5
thinc==7.4.5
threadpoolctl==3.1.0
tokenizers==0.10.3
toml==0.10.2
tomli==1.2.1
torch==1.9.0
tqdm==4.62.2
transformers==4.9.2
typed-ast==1.4.3
typer==0.3.2
typing-extensions==3.10.0.2
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.6
wasabi==0.8.2
wrapt==1.12.1
xgboost==1.4.2
zipp==3.10.0
